+++ Source/ACE.DatLoader/PackableExtensions.cs
@@ -103,7 +103,21 @@ namespace ACE.DatLoader
 
         /// <summary>
         /// A PackedHashTable uses a UInt16 for length, and a UInt16 for bucket size.
-        /// I don't know what the bucket size is actually usd for.
+        /// We don't need to worry about the bucket size with C#.
+        /// </summary>
+        public static void UnpackPackedHashTable(this Dictionary<uint, uint> value, BinaryReader reader)
+        {
+            var totalObjects = reader.ReadUInt16();
+            /*var bucketSize = */
+            reader.ReadUInt16();
+
+            for (int i = 0; i < totalObjects; i++)
+                value.Add(reader.ReadUInt32(), reader.ReadUInt32());
+        }
+
+        /// <summary>
+        /// A PackedHashTable uses a UInt16 for length, and a UInt16 for bucket size.
+        /// We don't need to worry about the bucket size with C#.
         /// </summary>
         public static void UnpackPackedHashTable<T>(this Dictionary<uint, T> value, BinaryReader reader) where T : IUnpackable, new()
         {